+++ NEWS.md
@@ -1,4 +1,4 @@
-# Require 1.1.0.9030 (development version)
+# Require 1.1.0.9037 (development version)
 
 ## enhancements